Extract low-level step/fixture and lambda fixtures to ExtendedApi
